% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseinput.R
\name{chooseinput}
\alias{chooseinput}
\title{chooseinput}
\usage{
chooseinput(country, asia_filter)
}
\arguments{
\item{country}{inputID parameter based on UI page}

\item{asia_filter}{filtered dataset inclding active cases}
}
\description{
This function is used to launch the covid_asia shiny app
}
