% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chosen_metric.R
\name{chosen_metric}
\alias{chosen_metric}
\title{chosen_metric}
\usage{
chosen_metric(data, input)
}
\arguments{
\item{data}{input data is the filtered dataset including active cases}

\item{input}{dataframe generated from shiny based on user control input}
}
\description{
This function is used to launch the covid_asia shiny app
}
