% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseinput.R
\name{chooseinput}
\alias{chooseinput}
\title{chooseinput}
\usage{
chooseinput(country, asia_filter)
}
\arguments{
\item{country}{inputID parameter based on UI page}

\item{asia_filter}{filtered dataset inclding active cases}
}
\description{
This function replaces the functionality of the built-in selectinput function of shiny based on the parameters mentioned in the ui part of the app.
}
